Update dkp.lua
Made a small patch in this code aswell. Whenever a manual edit is done to a members DKP, it will be presented in the officer chat, which was the most logical solution for us without spamming our guild chat about it. Also something that could be specified using options in the UI?